LayoutTests/canvas/philip/tests/2d.transformation.rotate.wrapnegative-expected.txt

< [index] >
2d.transformation.rotate.wrapnegative
rotate() wraps large negative values correctly
References: 2d.transformation.rotate
Actual output:
Expected output:

Passed
